requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt
#
annotated-types==0.6.0
    # via pydantic
attrs==23.1.0
    # via jsonschema
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.3.2
    # via requests
deepdiff==6.3.1
    # via acto (pyproject.toml)
exceptiongroup==1.2.0
    # via pytest
exrex==0.11.0
    # via acto (pyproject.toml)
google-auth==2.25.2
    # via kubernetes
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
jsonpatch==1.33
    # via acto (pyproject.toml)
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.17.3
    # via acto (pyproject.toml)
kubernetes==26.1.0
    # via acto (pyproject.toml)
numpy==1.26.2
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
ordered-set==4.1.0
    # via deepdiff
packaging==23.2
    # via pytest
pandas==2.0.3
    # via acto (pyproject.toml)
pluggy==1.4.0
    # via pytest
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.5.2
    # via acto (pyproject.toml)
pydantic-core==2.14.5
    # via pydantic
pyrsistent==0.20.0
    # via jsonschema
pytest==7.4.4
    # via acto (pyproject.toml)
python-dateutil==2.8.2
    # via
    #   kubernetes
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   acto (pyproject.toml)
    #   kubernetes
requests==2.31.0
    # via
    #   acto (pyproject.toml)
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.5
    # via acto (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
six==1.16.0
    # via
    #   kubernetes
    #   python-dateutil
tomli==2.0.1
    # via pytest
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via
    #   kubernetes
    #   requests
websocket-client==1.7.0
    # via kubernetes
gitpython==3.1.40
    # via gitpython
PyGithub==2.1.1
    # via PyGithub
python-gitlab==2.1
    # via python-gitlab
# The following packages are considered to be unsafe in a requirements file:
# setuptools
